app/requirements.txt
langchain
langchain-openai
openai
python-dotenv
langchain_community
fastapi
uvicorn
pydantic
python-docx